feature [fop]: Revamped functional operators with built-in bonding
Now `([[lv]:]op[:rv])` syntax allows for much more powerful functional
operators. They allow for built in bonding such as `(a:op:b)` being the
same as existing `a&(op)&b` syntax. This syntax is quite dense and
complex and therefore properly indenting it is important:

    (:-:@.len)

vs

    (: - : @.len)

is minor but can help calrity.

Signed-off-by: Vihan B <contact@vihan.org>